% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ganttChart.R
\name{ganttChart}
\alias{ganttChart}
\title{Plot a Gantt-Chart}
\usage{
ganttChart(
  excel_file,
  title = "Gantt-Chart",
  point_size = 5,
  date_breaks = "4 weeks",
  minor_breaks = "1 week",
  save = FALSE
)
}
\arguments{
\item{excel_file}{The file containing the tasks and the corresponding dates}

\item{title}{Title of the plot}

\item{point_size}{Size of the points plotted in the Gantt-Chart}

\item{save}{Save the plot as a .jpg file}
}
\value{

}
\description{
Plot a Gantt-Chart
}
